Restructure to put the solution first
Summarize our chosen solution at the top of the document and
restructure it to lift those sections to the top. Move the backround
and discussion down to an analysis section. Shorten the future work
section and remove discussion of building images inside Kubernetes,
since I wasn't sure about the motivation. (We can re-add that section
after discussion if I missed something.)